+++ sdk/core/azure-core-experimental/src/main/java/com/azure/core/experimental/util/polling/SyncOperationLocationPollingStrategy.java
@@ -16,27 +16,26 @@ import com.azure.core.http.rest.Response;
 import com.azure.core.util.BinaryData;
 import com.azure.core.util.logging.ClientLogger;
 import com.azure.core.util.polling.LongRunningOperationStatus;
-import com.azure.core.util.polling.OperationResourcePollingStrategy;
 import com.azure.core.util.polling.PollResponse;
 import com.azure.core.util.polling.PollingContext;
 import com.azure.core.util.polling.PollingStrategyOptions;
+import com.azure.core.util.polling.SyncOperationResourcePollingStrategy;
 import com.azure.core.util.serializer.ObjectSerializer;
 import com.azure.core.util.serializer.TypeReference;
-import reactor.core.publisher.Mono;
 
 import java.time.Duration;
 import java.time.OffsetDateTime;
 
 /**
- * Implements an operation location polling strategy, from Operation-Location.
+ * Implements a synchronous operation location polling strategy, from Operation-Location.
  *
  * @param <T> the type of the response type from a polling call, or BinaryData if raw response body should be kept
  * @param <U> the type of the final result object to deserialize into, or BinaryData if raw response body should be
  *           kept
  */
-public class OperationLocationPollingStrategy<T, U> extends OperationResourcePollingStrategy<T, U> {
+public class SyncOperationLocationPollingStrategy<T, U> extends SyncOperationResourcePollingStrategy<T, U> {
 
-    private static final ClientLogger LOGGER = new ClientLogger(OperationLocationPollingStrategy.class);
+    private static final ClientLogger LOGGER = new ClientLogger(SyncOperationLocationPollingStrategy.class);
 
     private static final HttpHeaderName OPERATION_LOCATION_HEADER
         = HttpHeaderName.fromString("Operation-Location");
@@ -53,7 +52,7 @@ public class OperationLocationPollingStrategy<T, U> extends OperationResourcePol
      * @param pollingStrategyOptions options to configure this polling strategy.
      * @throws NullPointerException if {@code pollingStrategyOptions} is null.
      */
-    public OperationLocationPollingStrategy(PollingStrategyOptions pollingStrategyOptions) {
+    public SyncOperationLocationPollingStrategy(PollingStrategyOptions pollingStrategyOptions) {
         super(OPERATION_LOCATION_HEADER, pollingStrategyOptions);
         this.endpoint = pollingStrategyOptions.getEndpoint();
         this.serializer = pollingStrategyOptions.getSerializer() != null
@@ -64,8 +63,8 @@ public class OperationLocationPollingStrategy<T, U> extends OperationResourcePol
      * {@inheritDoc}
      */
     @Override
-    public Mono<PollResponse<T>> onInitialResponse(Response<?> response, PollingContext<T> pollingContext,
-                                                   TypeReference<T> pollResponseType) {
+    public PollResponse<T> onInitialResponse(Response<?> response, PollingContext<T> pollingContext,
+                                             TypeReference<T> pollResponseType) {
         HttpHeader operationLocationHeader = response.getHeaders().get(OPERATION_LOCATION_HEADER);
         if (operationLocationHeader != null) {
             pollingContext.setData(OPERATION_LOCATION_HEADER.getCaseSensitiveName(),
@@ -79,41 +78,35 @@ public class OperationLocationPollingStrategy<T, U> extends OperationResourcePol
             || response.getStatusCode() == 202
             || response.getStatusCode() == 204) {
             final Duration retryAfter = ImplUtils.getRetryAfterFromHeaders(response.getHeaders(), OffsetDateTime::now);
-            final Mono<PollResponse<T>> pollResponseMono;
             if (HttpMethod.PUT.name().equalsIgnoreCase(httpMethod)
                 || HttpMethod.PATCH.name().equalsIgnoreCase(httpMethod)) {
                 // PUT has initial response body as resultType
                 // we expect Response<?> be either Response<BinaryData> or Response<U>
                 // if it is Response<T>, PollingUtils.serializeResponse would miss read-only properties
-                pollResponseMono = PollingUtils.serializeResponse(response.getValue(), serializer)
-                    .map(initialResponseBody -> {
-                        pollingContext.setData(
-                            PollingConstants.INITIAL_RESOURCE_RESPONSE_BODY, initialResponseBody.toString());
-                        return new PollResponse<>(LongRunningOperationStatus.IN_PROGRESS, null, retryAfter);
-                    });
+                BinaryData initialResponseBody = PollingUtils.serializeResponseSync(response.getValue(), serializer);
+                pollingContext.setData(PollingConstants.INITIAL_RESOURCE_RESPONSE_BODY, initialResponseBody.toString());
+                return new PollResponse<>(LongRunningOperationStatus.IN_PROGRESS, null, retryAfter);
             } else {
                 // same as OperationResourcePollingStrategy
-                pollResponseMono = PollingUtils.convertResponse(response.getValue(), serializer, pollResponseType)
-                    .map(value -> new PollResponse<>(LongRunningOperationStatus.IN_PROGRESS, value, retryAfter));
+                return new PollResponse<>(LongRunningOperationStatus.IN_PROGRESS,
+                    PollingUtils.convertResponseSync(response.getValue(), serializer, pollResponseType), retryAfter);
             }
-            return pollResponseMono.switchIfEmpty(Mono.fromSupplier(() -> new PollResponse<>(
-                LongRunningOperationStatus.IN_PROGRESS, null, retryAfter)));
-        } else {
-            return Mono.error(new AzureException(String.format("Operation failed or cancelled with status code %d,"
-                    + ", '%s' header: %s, and response body: %s", response.getStatusCode(), OPERATION_LOCATION_HEADER,
-                operationLocationHeader, PollingUtils.serializeResponse(response.getValue(), serializer))));
         }
+
+        throw LOGGER.logExceptionAsError(new AzureException(String.format(
+            "Operation failed or cancelled with status code %d, '%s' header: %s, and response body: %s",
+            response.getStatusCode(), OPERATION_LOCATION_HEADER, operationLocationHeader,
+            PollingUtils.serializeResponseSync(response.getValue(), serializer))));
     }
 
     /**
      * {@inheritDoc}
      */
-    @Override
-    public Mono<U> getResult(PollingContext<T> pollingContext, TypeReference<U> resultType) {
+    public U getResult(PollingContext<T> pollingContext, TypeReference<U> resultType) {
         if (pollingContext.getLatestResponse().getStatus() == LongRunningOperationStatus.FAILED) {
-            return Mono.error(new AzureException("Long running operation failed."));
+            throw LOGGER.logExceptionAsError(new AzureException("Long running operation failed."));
         } else if (pollingContext.getLatestResponse().getStatus() == LongRunningOperationStatus.USER_CANCELLED) {
-            return Mono.error(new AzureException("Long running operation cancelled."));
+            throw LOGGER.logExceptionAsError(new AzureException("Long running operation cancelled."));
         }
         String httpMethod = pollingContext.getData(PollingConstants.HTTP_METHOD);
         if (HttpMethod.PUT.name().equalsIgnoreCase(httpMethod)
@@ -121,22 +114,21 @@ public class OperationLocationPollingStrategy<T, U> extends OperationResourcePol
             // take the initial response body from PollingContext, and de-serialize as final result
             BinaryData initialResponseBody =
                 BinaryData.fromString(pollingContext.getData(PollingConstants.INITIAL_RESOURCE_RESPONSE_BODY));
-            return PollingUtils.deserializeResponse(initialResponseBody, serializer, resultType);
+            return PollingUtils.deserializeResponseSync(initialResponseBody, serializer, resultType);
         } else if (HttpMethod.POST.name().equalsIgnoreCase(httpMethod)) {
             // take the last poll response body from PollingContext,
             // and de-serialize the "result" property as final result
             BinaryData latestResponseBody =
                 BinaryData.fromString(pollingContext.getData(PollingConstants.POLL_RESPONSE_BODY));
-            return PollingUtils.deserializeResponse(latestResponseBody, serializer, POST_POLL_RESULT_TYPE_REFERENCE)
-                .flatMap(value -> {
-                    if (value.getResult() != null) {
-                        return PollingUtils.deserializeResponse(value.getResult(), serializer, resultType);
-                    } else {
-                        return Mono.error(new AzureException("Cannot get final result"));
-                    }
-                }).switchIfEmpty(Mono.error(new AzureException("Cannot get final result")));
+            PostPollResult pollResult =
+                PollingUtils.deserializeResponseSync(latestResponseBody, serializer, POST_POLL_RESULT_TYPE_REFERENCE);
+            if (pollResult != null && pollResult.getResult() != null) {
+                return PollingUtils.deserializeResponseSync(pollResult.getResult(), serializer, resultType);
+            } else {
+                throw LOGGER.logExceptionAsError(new AzureException("Cannot get final result"));
+            }
         } else {
-            return Mono.error(new AzureException("Cannot get final result"));
+            throw LOGGER.logExceptionAsError(new AzureException("Cannot get final result"));
         }
     }
 }