+++ sdk/core/azure-core-experimental/CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Features Added
 
-- Added `OperationLocationPollingStrategy` for the new polling strategy.
+- Added `OperationLocationPollingStrategy` and `SyncOperationLocationPollingStrategy` for the new polling strategy.
 - Added `PollResult` which represents a general poll result for new polling strategy.
 
 ### Breaking Changes